Refactoring that method and pushing it out to utils
